2.4.0 Update

+++ src/main/java/net/fneifnox/custommobattributes/mixin/damage/SnowballMixin.java
@@ -13,23 +13,17 @@ import static net.fneifnox.custommobattributes.CustomMobAttributes.CONFIG;
 
 @Mixin(SnowballEntity.class)
 public class SnowballMixin {
-    @Redirect(
-            method = "onEntityHit",
-            at = @At(
-                    value = "INVOKE",
-                    target = "Lnet/minecraft/entity/Entity;serverDamage(Lnet/minecraft/entity/damage/DamageSource;F)V"
-            )
-    )
+    @Redirect(method = "onEntityHit", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/Entity;serverDamage(Lnet/minecraft/entity/damage/DamageSource;F)V"))
      private void redirectDamage(Entity entity, DamageSource source, float originalDamage) {
-        float multiplier = 1f;
+        double multiplier = 1f;
         SnowballEntity snowball = (SnowballEntity)(Object)this;
         if (snowball.getOwner() instanceof SnowGolemEntity) {
             multiplier = CONFIG.damageMultiplierForSnowGolem() * CONFIG.damageMultiplierForAll();
         }
 
-        float finalDamage = originalDamage * multiplier;
+        double finalDamage = originalDamage * multiplier;
         if (entity.getWorld() instanceof ServerWorld serverWorld) {
-            entity.damage(serverWorld, source, finalDamage);
+            entity.damage(serverWorld, source, (float) finalDamage);
         }
     }
 }
